Created conftest.py
