+++ tests/api/store/test_store_happy_path.py
@@ -19,25 +19,6 @@ def test_get_pet_list_by_status():
         assert False, "Response is not in JSON format"
     assert isinstance(response, dict), "Expected response to be a dictionary"
 
-@pytest.fixture
-def new_order():
-    new_order = {
-        "id": 0,
-        "petId": 0,
-        "quantity": 1,
-        "shipDate": "2024-07-16T19:20:30.45Z",
-        "status": "placed",
-        "complete": True
-    }
-    response = requests.post(ORDER_ENDPOINT, json=new_order)
-    assert response.status_code == 200, f"Failed to create order, status code: {response.status_code}"
-
-    order = response.json()
-    yield order
-
-    delete_response = requests.delete(f"{ORDER_ENDPOINT}/{order['id']}")
-    assert delete_response.status_code == 200, f"Failed to delete order, status code: {delete_response.status_code}"
-
 
 def test_create_order(new_order):
     assert new_order["id"] is not None
@@ -51,23 +32,6 @@ def test_get_order_by_id(new_order):
     order = response.json()
     assert order["id"] == new_order["id"]
     assert order["status"] == new_order["status"] 
-
-
-@pytest.fixture
-def delete_order():
-    order = {
-        "id": 0,
-        "petId": 0,
-        "quantity": 1,
-        "shipDate": "2024-07-16T19:20:30.45Z",
-        "status": "placed",
-        "complete": True
-    }
-    response = requests.post(ORDER_ENDPOINT, json=order)
-    assert response.status_code == 200, f"Failed to create order, status code: {response.status_code}"
-
-    order = response.json()
-    yield order
 
 
 def test_delete_order(delete_order):
